+++ packages/studio-base/src/test/mocks/mockMessage.ts
@@ -4,6 +4,13 @@
 
 import { MessageEvent } from "@foxglove/studio";
 
+/**
+ * Helper function for generating mock messages for tests and stories.
+ *
+ * @param message the message body
+ * @param fields fields in the message object to override
+ * @returns a MessageEvent
+ */
 export function mockMessage<T>(message: T, fields?: Partial<MessageEvent<T>>): MessageEvent<T> {
   return {
     topic: "topic",
